# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Displays my resume main sections as APIs",
        "title": "Resume API",
        "contact": {
            "name": "flmailla",
            "email": "florent@maillard.icu"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/resume/v1",
    "paths": {
        "/experience/{experience_id}/skills": {
            "get": {
                "description": "Retrieve all the skills for a given experience",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills",
                    "Experience"
                ],
                "summary": "Get the experience skills",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Experience ID",
                        "name": "experience_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Skill"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    }
                }
            }
        },
        "/profiles/{profile_id}": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": []
                    }
                ],
                "description": "Retrieve the profil information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get a profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    }
                }
            }
        },
        "/profiles/{profile_id}/educations": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": []
                    }
                ],
                "description": "Retrieve all the education lines for a given profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Education",
                    "Profile"
                ],
                "summary": "Get a profile educations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Education"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    }
                }
            }
        },
        "/profiles/{profile_id}/experiences": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": []
                    }
                ],
                "description": "Retrieve all the experiences for a given profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experience",
                    "Profile"
                ],
                "summary": "Get a profile experiences",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Experience"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    }
                }
            }
        },
        "/profiles/{profile_id}/licences": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": []
                    }
                ],
                "description": "Retrieve all the Licences for a given profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Licence",
                    "Profile"
                ],
                "summary": "Get a profile Licences",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Licence"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    }
                }
            }
        },
        "/profiles/{profile_id}/skills": {
            "get": {
                "description": "Retrieve all the skills for a given profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills",
                    "Profile"
                ],
                "summary": "Get a profile skills",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Skill"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    }
                }
            }
        },
        "/skills": {
            "get": {
                "description": "Retrieve all the skills in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Get all the skills",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Skill"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponsNotFounde"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ErrorResponsNotFounde": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "error": {
                    "type": "string",
                    "example": "User not found"
                },
                "message": {
                    "type": "string",
                    "example": "The requested profile could not be found"
                }
            }
        },
        "models.Education": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "issued": {
                    "type": "string"
                },
                "profile": {
                    "$ref": "#/definitions/models.Profile"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Experience": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "location": {
                    "type": "string"
                },
                "profile": {
                    "$ref": "#/definitions/models.Profile"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Skill"
                    }
                },
                "startDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Licence": {
            "type": "object",
            "properties": {
                "expires": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "issuedAt": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "licenceType": {
                    "$ref": "#/definitions/models.LicenceType"
                },
                "profile": {
                    "$ref": "#/definitions/models.Profile"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.LicenceType": {
            "type": "string",
            "enum": [
                "Licnce",
                "Certification"
            ],
            "x-enum-varnames": [
                "LICENCE",
                "CERTIFICATION"
            ]
        },
        "models.Profile": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "birthDate": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "headline": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "lastName": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "integer",
                    "format": "int32"
                },
                "pronoun": {
                    "type": "string"
                }
            }
        },
        "models.Skill": {
            "description": "Skill information",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "git"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": "Grants read and write access to administrative information",
                "write": "Grants write access"
            }
        }
    }
}